Adding validation to query and if user exists or not
