+++ package.json
@@ -36,6 +36,9 @@
   },
   "dependencies": {
     "@prisma/client": "^4.0.0",
-    "express": "^4.18.1"
+    "express": "^4.18.1",
+    "express-validation": "^4.1.0",
+    "joi": "^17.6.0",
+    "winston": "^3.8.1"
   }
 }
