+++ src/common/validators/booking.validator.ts
@@ -1,0 +1,7 @@
+import Joi from "joi";
+
+export const userIdValidation = {
+  query: Joi.object({
+    userId: Joi.string().required(),
+  }),
+};
